+++ src/components/BibleComponents/VerseContainer/index.js
@@ -1,0 +1,35 @@
+import Card from "@mui/material/Card";
+import CardContent from "@mui/material/CardContent";
+import Typography from "@mui/material/Typography";
+
+import ShowVerse from "../ShowVerse";
+
+import { useContext } from "react";
+
+import Books from "../../../context/Books";
+
+function VerseContainer({ verse, title, handleClick }) {
+  const { books } = useContext(Books);
+
+  function findBook(bookid) {
+    if (books.find((book) => book.bookid === bookid) === undefined) {
+      return "livro não encontrado 😢";
+    }
+    return books.find((book) => book.bookid === bookid);
+  }
+
+  return (
+    <Card sx={{ margin: "50px auto", maxWidth: "80%" }} onCLick={handleClick}>
+      <CardContent>
+        {title && <Typography variant="h5">{title}</Typography>}
+        <Typography variant="h6">Livro:{findBook(verse.book).name}</Typography>
+        <Typography variant="p" sx={{ display: "block", margin: "5px" }}>
+          Capitulo:{verse.chapter}
+        </Typography>
+        <ShowVerse verse={verse} />
+      </CardContent>
+    </Card>
+  );
+}
+
+export default VerseContainer;
