+++ src/pages/Bible/index.js
@@ -1,7 +1,6 @@
 import BibleReader from "../../components/BibleComponents/BibleReader/index";
 import Loading from "../../components/GeneralComponents/Loading/index";
 import RandomVerse from "../../components/BibleComponents/RandomVerse/index";
-import NavBar from "../../components/GeneralComponents/NavBar";
 
 import CurrentBook from "../../context/CurrentBook";
 import Books from "../../context/Books";
